Курсовая работа (Бот в telegram).

+++ src/main/resources/liquibase/scripts/notification_task.sql
@@ -2,8 +2,8 @@
 
 -- changeset john:1
 CREATE TABLE notification_task (
-    id SERIAL PRIMARY KEY,
-    chat_id SERIAL NOT NULL,
+    id BIGINT PRIMARY KEY,
+    chat_id BIGINT NOT NULL,
     massage TEXT NOT NULL,
     right_date_time TIMESTAMP NOT NULL
 )
